+++ packages/server-wallet/src/objectives/objective-manager.ts
@@ -2,7 +2,7 @@ import {Logger} from 'pino';
 import {unreachable} from '@statechannels/wallet-core';
 
 import {Bytes32} from '../type-aliases';
-import {ChannelOpener} from '../protocols/open-channel';
+import {ChannelOpener} from '../protocols/channel-opener';
 import {ChannelCloser} from '../protocols/close-channel';
 import {Store} from '../wallet/store';
 import {ChainServiceInterface} from '../chain-service';
